python_bot/requirements.txt
aiogram
apscheduler
sqlalchemy
aiosqlite
anthropic
pydantic
pydantic-settings
python-dotenv
